+++ newton/__init__.py
@@ -34,7 +34,6 @@ from .geometry import (
     GEO_SDF,
     GEO_SPHERE,
     SDF,
-    Contacts,
     Mesh,
     create_box,
     create_capsule,
@@ -59,6 +58,7 @@ from .sim import (
     JOINT_PRISMATIC,
     JOINT_REVOLUTE,
     JOINT_UNIVERSAL,
+    Contacts,
     Control,
     Model,
     ModelBuilder,
