+++ newton/sim/__init__.py
@@ -16,6 +16,7 @@
 from .articulation import eval_fk, eval_ik
 from .builder import ModelBuilder
 from .collide import CollisionPipeline, count_rigid_contact_points
+from .contacts import Contacts
 from .control import Control
 from .flags import (
     NOTIFY_FLAG_BODY_INERTIAL_PROPERTIES,
@@ -64,6 +65,7 @@ __all__ = [
     "NOTIFY_FLAG_JOINT_PROPERTIES",
     "NOTIFY_FLAG_SHAPE_PROPERTIES",
     "CollisionPipeline",
+    "Contacts",
     "Control",
     "Model",
     "ModelBuilder",
