+++ newton/sim/contacts.py
